Implement displacement auto-search

+++ rendering/envs/target_env.py
@@ -35,6 +35,7 @@ class TargetEnvWrapper:
         robot_disp=None,
         unlimited=False,
         episode=0,
+        dry_run=False,
     ):
         data = np.load(os.path.join(source_robot_states_path, "source_robot_states", f"{episode}.npz"), allow_pickle=True)
         info = ROBOT_CAMERA_POSES_DICT[robot_dataset]
@@ -71,14 +72,16 @@ class TargetEnvWrapper:
         gripper_width_list = []
         success = True
 
-        mask_dir = Path(save_paired_images_folder_path) / f"{self.target_name}_replay_mask"
-        video_dir = Path(save_paired_images_folder_path) / f"{self.target_name}_replay_video"
-        mask_dir.mkdir(parents=True, exist_ok=True)
-        video_dir.mkdir(parents=True, exist_ok=True)
-        mask_path  = mask_dir / f"{episode}.mp4"
-        video_path = video_dir / f"{episode}.mp4"
         mask_frames = []
         video_frames = []
+        mask_path = video_path = None
+        if not dry_run:
+            mask_dir = Path(save_paired_images_folder_path) / f"{self.target_name}_replay_mask"
+            video_dir = Path(save_paired_images_folder_path) / f"{self.target_name}_replay_video"
+            mask_dir.mkdir(parents=True, exist_ok=True)
+            video_dir.mkdir(parents=True, exist_ok=True)
+            mask_path  = mask_dir / f"{episode}.mp4"
+            video_path = video_dir / f"{episode}.mp4"
         suggestion = np.zeros(3)
         for pose_index in range(num_robot_poses):
             target_pose=target_pose_array[pose_index].copy()
@@ -151,25 +154,31 @@ class TargetEnvWrapper:
             joint_angles_list.append(joint_angles)
 
 
-            target_robot_img, target_robot_seg_img = self.target_env.get_observation_fast(white_background=True, width=self.camera_width, height=self.camera_height)
-            mask_frames.append(target_robot_seg_img) 
-            video_frames.append(target_robot_img)
+            target_robot_img, target_robot_seg_img = self.target_env.get_observation_fast(
+                white_background=True,
+                width=self.camera_width,
+                height=self.camera_height,
+            )
+            if not dry_run:
+                mask_frames.append(target_robot_seg_img)
+                video_frames.append(target_robot_img)
             # cv2.imwrite(os.path.join(save_paired_images_folder_path, f"{self.target_name}_rgb", f"{episode}/{pose_index}.png"), cv2.cvtColor(target_robot_img, cv2.COLOR_RGB2BGR))
             # cv2.imwrite(os.path.join(save_paired_images_folder_path, f"{self.target_name}_mask", f"{episode}/{pose_index}.png"), target_robot_seg_img * 255)
-        mask_frames_np = np.stack(mask_frames, axis=0).astype(np.uint8) * 255
-        video_frames_np = np.stack(video_frames, axis=0)
-        iio.imwrite(
-            mask_path,
-            mask_frames_np,
-            fps   = 30,
-            codec = "libx264"
-        )
-        iio.imwrite(
-            video_path,
-            video_frames_np,
-            fps   = 30,
-            codec = "libx264"
-        )
+        if not dry_run:
+            mask_frames_np = np.stack(mask_frames, axis=0).astype(np.uint8) * 255
+            video_frames_np = np.stack(video_frames, axis=0)
+            iio.imwrite(
+                mask_path,
+                mask_frames_np,
+                fps=30,
+                codec="libx264",
+            )
+            iio.imwrite(
+                video_path,
+                video_frames_np,
+                fps=30,
+                codec="libx264",
+            )
         
         if success:
             if unlimited == False:
@@ -188,26 +197,40 @@ class TargetEnvWrapper:
             # np.save(joint_angles_npy_path, joint_angles_array)
             # offset_npy_path = os.path.join(save_paired_images_folder_path, "source_robot_states", f"{self.target_name}", "offsets", f"{episode}.npy")
             # np.save(offset_npy_path, robot_disp)
-            os.makedirs(os.path.join(save_paired_images_folder_path, "target_robot_states", f"{self.target_name}"), exist_ok=True)
-            target_pose_array   = np.vstack(target_pose_list)
-            joint_angles_array  = np.vstack(joint_angles_list)
-            gripper_width_array = np.asarray(gripper_width_list)
-            offset_array        = robot_disp
-            state_npz_path = os.path.join(
-                save_paired_images_folder_path,
-                "target_robot_states",
-                f"{self.target_name}",
-                f"{episode}.npz",
+            if not dry_run:
+                os.makedirs(
+                    os.path.join(
+                        save_paired_images_folder_path,
+                        "target_robot_states",
+                        f"{self.target_name}",
+                    ),
+                    exist_ok=True,
+                )
+                target_pose_array = np.vstack(target_pose_list)
+                joint_angles_array = np.vstack(joint_angles_list)
+                gripper_width_array = np.asarray(gripper_width_list)
+                offset_array = robot_disp
+                state_npz_path = os.path.join(
+                    save_paired_images_folder_path,
+                    "target_robot_states",
+                    f"{self.target_name}",
+                    f"{episode}.npz",
+                )
+                np.savez(
+                    state_npz_path,
+                    target_pose=target_pose_array,
+                    joint_angles=joint_angles_array,
+                    gripper_width=gripper_width_array,
+                    offsets=offset_array,
+                )
+            steps = len(target_pose_list)
+            logger.debug(
+                "Episode %s success with displacement %s, steps %s",
+                episode,
+                robot_disp,
+                steps,
             )
-            np.savez(
-                state_npz_path,
-                target_pose   = target_pose_array,
-                joint_angles  = joint_angles_array,
-                gripper_width = gripper_width_array,
-                offsets       = offset_array,
-            )
-            logger.debug("Episode %s success with displacement %s", episode, robot_disp)
-            return success, suggestion
+            return success, suggestion, steps
         else:
             print(f"\033[91m[FAILURE] Could not reach target pose for {self.target_name} – episode {episode}\033[0m")
             logger.debug(
@@ -216,4 +239,5 @@ class TargetEnvWrapper:
                 robot_disp,
                 suggestion,
             )
-            return False, suggestion
+            steps = len(target_pose_list)
+            return False, suggestion, steps
